Front End & Back End : Fake database generator update
Now, when the server start, if there is no database, the server will inject :
- 130 users, 
- 1 363 posts,
- more than 6 500 like/dislike for posts,
-  10 270 comments,
- more than 50 000 like/dislike for comments.

Posts and comments have the attribute "createdAt", dates ranging from July 2021 to December 2022

When the front end requests posts or comments from the server, the API will only return posts where the creation date has passed. This gives the user the impression that there is activity on the site and thus encourages them to post.